Exact ayaka-plugin.
