+++ utils/ayaka-plugin/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "ayaka-plugin"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+ayaka-bindings-types = { path = "../ayaka-bindings-types" }
+ayaka-script = { path = "../ayaka-script" }
+serde = { version = "1.0", features = ["derive"] }
+anyhow = "1.0"
+rmp-serde = "1.1"
