+++ utils/ayaka-runtime/src/lib.rs
@@ -10,7 +10,6 @@
 #![feature(fn_traits)]
 #![feature(generators)]
 #![feature(once_cell)]
-#![feature(tuple_trait)]
 #![feature(unboxed_closures)]
 
 mod config;
